requirements.txt
# This file is autogenerated by Prequ.  To update, run:
#
#   prequ update
#
--find-links wheels

django-environ==0.4.5
mongo-to-sql==0.0.1
psycopg2-binary==2.8.5
pymongo==3.11.2
